backend/requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# Environment Management
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.7

# Security
django-ratelimit==4.1.0
django-csp==3.7

# File Processing
Pillow==10.0.1
python-magic==0.4.27

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.0

# AI and NLP
spacy==3.7.2
nltk==3.8.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
